Introduce an private list in the Plan structure. It allows extensions to store an planning info for using during (or after) execution.
Introduce the AQOPlanNode structure. Move AQO-related fields of the
Plan structure in this structure. Add this structure as an element
to the Plan->private field.

We allow copying of AQOPlanNode and disallow _read... and _out...
functions because it have sense and it needs to teach the Core to
serialize/deserialize RestrictInfo and all descendants of this struct.

Also, for simplifying the Core  patch:
1. replace parallel_workers field with parallel_divisor
2. move IsParallelTuplesProcessing() from the Core into the AQO code.

This commit changes the Core patch.