Allow AQO teaching on a plan with 'never executed' nodes.
Previously we suppressed such plans because 'never executed' node
is a kind of NULL value. We don't really know how many tuples (and
work) it require in the situation, when it will be executed even once.

Current commit is attempt to teach on such plans, may be partly.
We allow teaching on all nodes, that were executed. If node was
'never executed' and no any information we have in ML-knowledge base
for this feaure subspace - we add knowledge that this node can
produce one tuple.