Add into unsupported.sql demo for prediction of filtered tuples.
AQO need to predict total fetched tuples in a table. At a non-leaf
node we have prediction about input tuples - is a number of predicted
output rows in underlying node. But for Scan nodes we don't have any
prediction on number of fetched tuples.

So, if selectivity was wrong we could make bad choice of Scan operation.
For example, we could choose suboptimal index.