Don't predict and learn on plans without any underlying plane tables. Optimizer can do his work the best way without an AQO assistance. So, we excluded zero relids from a relids list.
Improve the pgbench.pl TAP test of AQO extension and check correctness
of cleaning procedure after deletion of all pgbench-related tables.

TODO: here we found a possible deadlock on the AQO ignorance table
creation. We have to solve ths problem or remove this machinery at all.
Suppress the source of this problem in the test.