Performance commit.
Teach add_query_text to check existence of record (or concurrent
insertion) and skip insertion in such case.
It allows us to make simple modifications in the code. Now, we skip
access to database in the DISABLED mode if it is not really
necessary.
Also, reduce a scope of custom lock on a query class in the aqo_planner
and executor_end hooks.