Add smart statement timeout for learning AQO.
AQO evaluates from the current aqo.statement_timeout value whether it has enough time to train.
If not, then it increases this time exponentially.
We can evaluate whether there was enough time to train the AQO by the average integral error.
If the internal error has not practically changed compared to the previous iteration, then the training time increases.
If the user needs an upper limit on query execution time, they can set a vanilla statement timeout.
The aqo.statement_timeout value will be stored in shmem, next to the training data. Initially it = 0 (default value).
The user has the ability to specify the GUC aqo.statement_timeout and terminates the request by analogy with statement timeout.