Add smart statement timeout for learning AQO.
AQO evaluates to have an enough time for training by the average integral error.
If the integral error hasn't changed for comparing with error from the previous iteration,
we increase the training time exponentially.
Start value of smart statement timeout as equal as aqo_statement timeout or 0.
If the user needs an upper limit on query execution time, he can set a vanilla statement timeout
setting aqo_statement_timeout as 0. Smart statement timeout in this option is unenabled.
The aqo.statement_timeout value is stored in aqo_queries. Initially it equals as 0 (default value).
The user has the ability to specify the GUC aqo.statement_timeout and
terminates the request by analogy with statement timeout.