Add compute_query_id parameter in aqo configure with value as regress. It is necessary for avoiding output Query Identifier while vanille's test are running.
(Look at more in explain.c:612.
We can get fail condition if only query identifier is not null)

Where clause 'NOT LIKE '%Query Identifier%'' is throwed away due to being necessary any more.
This addition parameter is appeared if we set compute_query_id parameter with value as 'auto'.
Appearance of the parameter is checked in only gucs test.

a.lepikhov: cherry-picked this commit down to stable13 just to have as
synchronized as possible versions of the branches.