Add a github actions script to test each pushed commit on a self hosted runner.
Branch name is a key to define the name of suitable PostgreSQL core branch:
use "stable[XX]" phrase in the name of git branch to trigger compiling and
launch of this commit with REL_[XX]_STABLE branch of the core.
If the branch name don't contain such phrase, use master branch.