Add couple of github actions flows on each push event: - run make installcheck over an instance in different modes. - run JOB benchmark [1] on a self hosted runner.
Utility scripts stores in the .github folder.

Branch name is a key to define the name of suitable PostgreSQL core branch:
use "stable[XX]" phrase in the name of git branch to trigger compiling and
launch of this commit with REL_[XX]_STABLE branch of the core.
If the branch name doesn't contain such a phrase, use master branch.

TODO:
=====

1. Add 'long' JOB test (parallel strategy disabled).
2. Add JOB test which would be executed up to full convergency of learning on
   each query.
3. Add installchecks with reusage of existed database and the AQO extension
   installed (sanity checks will be definitely broken but still).
4. Additional queries [2] can be a marker for successful learning.

[1] https://github.com/danolivo/jo-bench
[2] https://github.com/RyanMarcus/imdb_pg_dataset